Rename some BinaryPrimitives files (and move some functions)
Zero code changes... just renamed some files, moved the contents of one file to another, and deleted a then empty file.
